backend/ta_connect/requirements.txt
Django==6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
django-csp==4.0
python-decouple==3.8
psycopg[binary]==3.2.10
requests==2.32.4
setuptools
wheel
drf_yasg==1.21.11
django-ratelimit==4.1.0
python-dateutil
django-encrypted-model-fields
pytz
django-webpush